Split Response mapping
