+++ browser/mapping.go
@@ -49,40 +49,6 @@ func parseFrameClickOptions(
 	return copts, nil
 }
 
-// mapResponse to the JS module.
-func mapResponse(vu moduleVU, r *common.Response) mapping {
-	if r == nil {
-		return nil
-	}
-	rt := vu.Runtime()
-	maps := mapping{
-		"allHeaders": r.AllHeaders,
-		"body":       r.Body,
-		"frame": func() *goja.Object {
-			mf := mapFrame(vu, r.Frame())
-			return rt.ToValue(mf).ToObject(rt)
-		},
-		"headerValue":  r.HeaderValue,
-		"headerValues": r.HeaderValues,
-		"headers":      r.Headers,
-		"headersArray": r.HeadersArray,
-		"json":         r.JSON,
-		"ok":           r.Ok,
-		"request": func() *goja.Object {
-			mr := mapRequest(vu, r.Request())
-			return rt.ToValue(mr).ToObject(rt)
-		},
-		"securityDetails": r.SecurityDetails,
-		"serverAddr":      r.ServerAddr,
-		"size":            r.Size,
-		"status":          r.Status,
-		"statusText":      r.StatusText,
-		"url":             r.URL,
-	}
-
-	return maps
-}
-
 // mapJSHandle to the JS module.
 func mapJSHandle(vu moduleVU, jsh common.JSHandleAPI) mapping {
 	rt := vu.Runtime()
